refactor: move more classes
