+++ packages/core-types/package.json
@@ -51,6 +51,7 @@
     "dependencies": {
         "@js-soft/logging-abstractions": "^1.0.1",
         "@js-soft/ts-serval": "2.0.11",
+        "@nmshd/crypto": "^2.1.0",
         "json-stringify-safe": "^5.0.1",
         "luxon": "^3.5.0"
     },
