+++ packages/core-types/src/index.ts
@@ -2,6 +2,7 @@ export * from "./CoreAddress";
 export * from "./CoreDate";
 export * from "./CoreError";
 export * from "./CoreId";
-export * from "./ICoreIdHelper";
+export * from "./CoreIdHelper";
+export * from "./Random";
 export * from "./references";
 export * from "./SharedPasswordProtection";
