+++ packages/core-types/src/references/FileReference.ts
@@ -1,4 +1,5 @@
 import { type } from "@js-soft/ts-serval";
+import { CoreIdHelper } from "../CoreIdHelper";
 import { IReference, Reference } from "./Reference";
 
 export interface IFileReference extends IReference {}
@@ -6,7 +7,7 @@ export interface IFileReference extends IReference {}
 @type("FileReference")
 export class FileReference extends Reference implements IFileReference {
     protected static override preFrom(value: any): any {
-        // super.validateId(value, BackboneIds.file);
+        super.validateId(value, new CoreIdHelper("FIL", true));
 
         return value;
     }
