+++ packages/transport/src/util/index.ts
@@ -1,2 +1,1 @@
 export * from "./PasswordGenerator";
-export * from "./Random";
